Migrate from SQLite to LiteDB
Replaced SQLite and Entity Framework with LiteDB for improved performance and simplified data handling. Removed all SQLite-related code, including Entity Framework migrations and design-time DbContext factory. Refactored data services to use LiteDB collections and added manual relationship loading methods.

Introduced a new `Migration` project to migrate data from SQLite to LiteDB. Updated domain models with LiteDB attributes (`BsonId`, `BsonIgnore`) and added a `Sale` model for sales records. Updated commands, ViewModels, and interfaces to align with the new data layer.

Simplified configuration files (`appsettings.json`, `App.config`) and updated solution and package files to include LiteDB dependencies. Removed unused code and applied minor formatting and cleanup throughout the codebase.

+++ Mhyrenz_Interface/Commands/AddCommand.cs
@@ -64,7 +64,7 @@ namespace Mhyrenz_Interface.Commands
             var product = await _productService.Add(new Product
             {
                 Name = _viewModel.Name,
-                RetailPrice = (decimal)_viewModel.Price, // RESOLVE DECIMAL TO DOUBLE
+                RetailPrice = _viewModel.Price,
                 Qty = _viewModel.Qty,
                 CategoryId = _viewModel.SelectedCategory.Id,
                 Expiry = _viewModel.Expiry,
